[OLD] Encapsulate/requirements.txt
flet
pyinstaller
setuptools
python-dotenv